+++ src/error.rs
@@ -302,6 +302,9 @@ pub(crate) enum CorruptKind {
     /// An inode's file type is invalid.
     InodeFileType { inode: InodeIndex, mode: InodeMode },
 
+    /// An inode's timestamp is invalid.
+    InodeTimestamp { inode: InodeIndex },
+
     /// The target of a symlink is not a valid path.
     SymlinkTarget(InodeIndex),
 
@@ -489,6 +492,9 @@ impl Display for CorruptKind {
                     "inode {inode} has invalid file type: mode=0x{mode:04x}",
                     mode = mode.bits()
                 )
+            }
+            Self::InodeTimestamp { inode } => {
+                write!(f, "inode {inode} has invalid timestamp",)
             }
             Self::SymlinkTarget(inode) => {
                 write!(f, "inode {inode} has an invalid symlink path")