+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   and `Path`) to wrap the string in quotes, matching `std` behavior`.
 * Changed the `Debug` impl for `DirEntry` to just show the path,
   matching `std` behavior.
+* Added support for inode timestamps with a new `Timestamp` type.
 
 ## 0.9.3